marketing_optin/pyproject.toml
[tool.poetry]
name = "crewai_marketing_optin"
version = "0.1.0"
description = "A CrewAI project to identify customers not opted in for marketing"
authors = ["POSInnovator"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = ">=0.1.24"
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"
langchain-openai = "^0.3.5"
langchain = "^0.3.18"
langchain-community = "^0.3.17"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
